backend/instagram_utility/application_instagram_api/serializers.py
from rest_framework import serializers
from .models import InstagramModel


class InstagramSerializer(serializers.ModelSerializer):

    class Meta:
        model = InstagramModel
        fields = ('id', 'username', 'password')
