backend/requirements.txt
decorator==4.0.11
Django==2.0.3
django-cors-headers==2.2.0
djangorestframework==3.7.7
imageio==2.1.2
InstagramAPI==1.0.2
moviepy==0.2.3.2
numpy==1.14.1
Pillow==5.0.0
pytz==2018.3
requests==2.11.1
requests-toolbelt==0.7.0
tqdm==4.11.2
